+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Version 0.7.2 (TBA)
+# Version 0.7.2 (08.10.2018)
 * New Feature: Introduced funding currency symbols
 * Improvement: API key is now an optional
 * Bugfix: Ensure channel unsubscribe handler is registered before unsubscribe is called